+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/FlavourContext.java
@@ -1,10 +1,13 @@
 package org.fugerit.java.doc.project.facade;
 
 import lombok.*;
+import lombok.extern.slf4j.Slf4j;
+import org.fugerit.java.core.lang.helpers.StringUtils;
 
 import java.io.File;
 import java.util.List;
 
+@Slf4j
 @ToString
 @RequiredArgsConstructor
 public class FlavourContext {
@@ -32,6 +35,9 @@ public class FlavourContext {
 
     @Getter @Setter
     private boolean addLombok;
+
+    @Getter @Setter
+    private String flavourVersion;
 
     private String toClassName( String base, String splitString ) {
         StringBuilder buf = new StringBuilder();